minor cleanup - unused component
