+++ libs/application/templates/operating-license/src/fields/PaymentCharge/index.tsx
@@ -5,7 +5,6 @@ import { Box } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
 import { m } from '../../lib/messages'
 import { getValueViaPath } from '@island.is/application/core'
-import { Operation } from '../../lib/constants'
 import { getChargeItemCode, getCurrencyString } from '../../lib/utils'
 import { PaymentCatalogItem } from '@island.is/api/schema'
 import { useFormContext } from 'react-hook-form'
