+++ libs/application/templates/operating-license/src/fields/index.ts
@@ -1,7 +1,6 @@
 export { PaymentPending } from './PaymentPending'
 export { PaymentCharge } from './PaymentCharge'
 export { PropertyRepeater } from './PropertyRepeater'
-export { Bullets } from './Bullets'
 export { FileOverview } from './FileOverview'
 export { PropertyOverviewRepeater } from './PropertyOverviewRepeater'
-export { Congratulations } from './Done/Congratulations'
+export { Congratulations } from './Done'
